Fix usersync
