+++ internal/user/usersql/users.sql.go
@@ -9,19 +9,6 @@ import (
 	"github.com/google/uuid"
 )
 
-const assignGlobalAdmin = `-- name: AssignGlobalAdmin :exec
-UPDATE users
-SET
-	admin = TRUE
-WHERE
-	id = $1
-`
-
-func (q *Queries) AssignGlobalAdmin(ctx context.Context, id uuid.UUID) error {
-	_, err := q.db.Exec(ctx, assignGlobalAdmin, id)
-	return err
-}
-
 const count = `-- name: Count :one
 SELECT
 	COUNT(*)
@@ -302,19 +289,6 @@ func (q *Queries) ListMemberships(ctx context.Context, arg ListMembershipsParams
 		return nil, err
 	}
 	return items, nil
-}
-
-const revokeGlobalAdmin = `-- name: RevokeGlobalAdmin :exec
-UPDATE users
-SET
-	admin = FALSE
-WHERE
-	id = $1
-`
-
-func (q *Queries) RevokeGlobalAdmin(ctx context.Context, id uuid.UUID) error {
-	_, err := q.db.Exec(ctx, revokeGlobalAdmin, id)
-	return err
 }
 
 const update = `-- name: Update :one
